requirements.txt
numpy==1.23.4
pyyaml==6.0
tqdm==4.64.1
pillow==9.3.0
mediapy==1.1.4
scipy==1.9.3
mujoco==3.0.0
pytransform3d==3.4.0